+++ jjava-distro/src/main/java/org/dflib/jjava/distro/JJava.java
@@ -72,6 +72,7 @@ public class JJava {
 
                 .build();
 
+        kernel.onStartup();
         kernel.becomeHandlerForConnection(connection);
 
         connection.connect();