+++ RELEASE-NOTES.md
@@ -16,6 +16,7 @@
 - #86 ineffecient / incorrect classpath handling
 - #89 Deprecate redundant magics: %jars, %addMavenDependency
 - #90 Deprecate Ivy artifact syntax for %maven magic
+- #93 Extensions / kernel lifecycle
 
 ## 1.0-a5